Reworking brokers to allow for scheduling in redis

+++ ironworker-core/src/task/async.rs
@@ -76,7 +76,7 @@ macro_rules! impl_async_task_function {
                 self
             }
 
-            fn retries(mut self, count: usize) -> Self {
+            fn retries(mut self, count: u64) -> Self {
                 self.config.retries.replace(count);
                 self
             }
